+++ Sources/Protocols/MessagesLayoutDelegate.swift
@@ -52,12 +52,11 @@ public protocol MessagesLayoutDelegate: AnyObject {
     /// Specifies the size to use for a typing indicator view.
     ///
     /// - Parameters:
-    ///   - messagesCollectionView: The `MessagesCollectionView` in which this view will be displayed.
+    ///   - layout: The `MessagesCollectionViewFlowLayout` layout.
     ///
     /// - Note:
-    ///   The default value returned by this method is the width of the `messagesCollectionView` and
-    ///   a height of 52.
-    func typingIndicatorViewSize(in messagesCollectionView: MessagesCollectionView) -> CGSize
+    ///   The default value returned by this method is the width of the `messagesCollectionView` minus insets and a height of 62.
+    func typingIndicatorViewSize(for layout: MessagesCollectionViewFlowLayout) -> CGSize
 
     /// Specifies the top inset to use for a typing indicator view.
     ///
@@ -134,8 +133,11 @@ public extension MessagesLayoutDelegate {
         return .zero
     }
 
-    func typingIndicatorViewSize(in messagesCollectionView: MessagesCollectionView) -> CGSize {
-        return CGSize(width: messagesCollectionView.bounds.width, height: 48)
+    func typingIndicatorViewSize(for layout: MessagesCollectionViewFlowLayout) -> CGSize {
+        let collectionViewWidth = layout.messagesCollectionView.bounds.width
+        let contentInset = layout.messagesCollectionView.contentInset
+        let inset = layout.sectionInset.horizontal + contentInset.horizontal
+        return CGSize(width: collectionViewWidth - inset, height: 62)
     }
 
     func typingIndicatorViewTopInset(in messagesCollectionView: MessagesCollectionView) -> CGFloat {